requirements.txt
autopep8>=1.4.4
bandit>=1.6.2
beautifulsoup4~=4.8.2
bs4>=0.0.1
click>=7.0
colorama==0.4.1
configparser>=5.0.0
dateparser>=0.7.2
demisto-py>=2.0.11
Deprecated~=1.2.7
dictor>=0.1.5
docker>=4.2.0
flake8>=3.8.3
flatten_dict>=0.3.0
gitpython>=3.1.0
isort>=4.3.21
jinja2>=2.11.1
mergedeep>=1.3.0
mypy>=0.761
networkx>=2.4
packaging~=20.1
pathlib~=1.0.1
pipenv==2018.11.26
pykwalify>=1.7.0
PyPDF2>=1.26.0
pytest~=5.3.5
python-dateutil~=2.8.1
PyYAML>=5.3.1
requests>=2.22.0
ruamel.yaml>=0.16.5

setuptools~=41.2.0
tabulate>=0.8.7
tabulate>=0.8.7
ujson>=3.0.0
urllib3>=1.25.7
vulture>=1.3
wcmatch>=5.1
yamlordereddictloader>=0.4.0
